app: cancel balance request
